requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0